Use MSB as tag flag and improve tag management
This patch improves the previous encoding scheme, which embedded the object ID in the upper two bytes of a pointer. We now use bit 63 (the MSB) as the taint flag and reorganize the taint/tag management logic around this representation. This change eliminates sign-extension issues, and provides a cleaner and more robust tagging model.

Signed-off-by: Adel Belkhiri <adel.belkhiri@gmail.com>